api: use axios + react-query instead of redux
All we need right now is a server-state cache, so these libraries give
it to us very concisely and without boilerplate. What little
app-specific data we have can be stored in react context for now. We
should think about tuning the caching parameters for react-query and
storing state on disk for persistent/offline access next.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>